Updates on ChildIssues problem
