+++ src/components/EpicDetailView/helpers/ChildIssues.tsx
@@ -1,34 +1,13 @@
-import {
-  Box,
-  Button,
-  Center,
-  Divider,
-  Flex,
-  Loader,
-  ScrollArea,
-  Stack,
-  Text,
-  Title,
-} from "@mantine/core"
-import { useQuery } from "@tanstack/react-query"
-import { useEffect, useState } from "react"
+import { Box, Button, Divider, Flex, ScrollArea, Stack } from "@mantine/core"
+import { useState } from "react"
 import { DragDropContext } from "react-beautiful-dnd"
-import { useNavigate } from "react-router-dom"
 import { Issue } from "types"
-import { useCanvasStore } from "../../../lib/Store"
 import { CreateIssueModal } from "../../CreateIssue/CreateIssueModal"
-import { sortIssuesByRank } from "../../BacklogView/helpers/backlogHelpers"
 import { onDragEnd } from "../../BacklogView/helpers/draggingHelpers"
-import { getBacklogIssues } from "../../BacklogView/helpers/queryFetchers"
-import { resizeDivider } from "../../BacklogView/helpers/resizeDivider"
 import { ChildIssueWrapper } from "./ChildIssueWrapper"
 
-export function ChildIssues({ summary }: { summary: string }) {
-  const navigate = useNavigate()
+export function ChildIssues({ issues }: { issues: Issue[] }) {
   const [createIssueModalOpened, setCreateIssueModalOpened] = useState(false)
-  const projectKey = useCanvasStore((state) => state.selectedProject?.key)
-  const boardIds = useCanvasStore((state) => state.selectedProjectBoardIds)
-  const currentBoardId = boardIds[0]
 
   const [issuesWrappers, setIssuesWrappers] = useState(
     new Map<string, { issues: Issue[] }>()
@@ -41,55 +20,6 @@ export function ChildIssues({ summary }: { summary: string }) {
     setSearchedissuesWrappers((map) => new Map(map.set(key, value)))
   }
 
-  const { isLoading: isLoadingBacklogIssues, isError: isErrorBacklogIssues } =
-    useQuery({
-      queryKey: ["issues", projectKey, currentBoardId],
-      queryFn: () => getBacklogIssues(projectKey, currentBoardId),
-      enabled: !!projectKey,
-      onSuccess: (backlogIssues) => {
-        updateIssuesWrapper("Backlog", {
-          issues:
-            backlogIssues && backlogIssues instanceof Array
-              ? backlogIssues
-                  .filter((issue: Issue) => issue.epic === summary)
-                  .sort((issueA: Issue, issueB: Issue) =>
-                    sortIssuesByRank(issueA, issueB)
-                  )
-              : [],
-        })
-      },
-    })
-
-  useEffect(() => {
-    resizeDivider()
-  }, [isLoadingBacklogIssues])
-
-  if (isErrorBacklogIssues)
-    return (
-      <Center style={{ width: "100%", height: "100%" }}>
-        <Text w="300">
-          An error has occurred while loading. This is due to an internal error.
-          Please report this behavior to the developers. <br />
-        </Text>
-      </Center>
-    )
-
-  if (isLoadingBacklogIssues)
-    return (
-      <Center style={{ width: "100%", height: "100%" }}>
-        {projectKey ? (
-          <Loader />
-        ) : (
-          <Stack align="center">
-            <Title>No Project has been selected!</Title>
-            <Text>
-              Please go back to the Projects View section and select a project
-            </Text>
-            <Button onClick={() => navigate("/projectsview")}>Go back</Button>
-          </Stack>
-        )}
-      </Center>
-    )
   return (
     <Stack
       sx={{
@@ -117,10 +47,7 @@ export function ChildIssues({ summary }: { summary: string }) {
           >
             {searchedissuesWrappers.get("Backlog") && (
               <Box mr="md">
-                <ChildIssueWrapper
-                  id="Backlog"
-                  issues={searchedissuesWrappers.get("Backlog")!.issues}
-                />
+                <ChildIssueWrapper id="Backlog" issues={issues} />
               </Box>
             )}
             <Box mr="xs">
